requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==5.12.4
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.3.0
tenacity==8.2.3
toml==0.10.2

st-gsheets-connection
datetime
plotly.express
xlsxwriter
st-gsheets-connection==0.1.0
streamlit>=1.22.0
pandasql
datetime